Android: Bug beseitigen

+++ sockettest/AndroidUDPTCPTest/src/com/legocar/testingversion/utils/ServoCommunicationUtils.java
@@ -4,8 +4,8 @@ import android.util.Log;
 
 public class ServoCommunicationUtils {
 
-	public static String getServoControlMessage(int servo, int middleValue, int offset, float percent){
-		int value = middleValue+(int)((float)offset*percent/100.0);
+	public static String getServoControlMessage(int servo, int middleValue, int offset, double percent){
+		int value = middleValue+(int)(offset*percent/100.0);
 		Log.i("Message", value+"");
 		return "servo set "+servo+" "+value+"\n";
 	}
